Set 64bit variable size; Added size checks

+++ src/hexdump.h
@@ -4,6 +4,6 @@
 #define FALSE 0
 #define TRUE  1
 
-int print_hex(char *file_name,unsigned long int start_byte,unsigned long int no_of_bytes);
+int print_hex(char *file_name,unsigned long long int start_byte,unsigned long long int no_of_bytes);
 
 #endif
